refactor NPQ statements to use central calculator
