+++ app/components/finance/npq/payment_overviews/course.html.erb
@@ -1,0 +1,53 @@
+<section class="app-application__card">
+  <h2 class="govuk-heading-s"><%= t("#{ contract.course_identifier }", scope: "npq_courses") %></h2>
+
+  <header class="app-application-card__header">
+    <% calculator.milestones.order(name: :asc).each do |milestone| %>
+      <span class="govuk-body-s">
+        <%= milestone.declaration_type.titleize %>
+        <br>
+        <%= calculator.total_participants_for(milestone) %>
+      </span>
+    <% end %>
+
+    <span class="govuk-body-s"><%= t("finance.total_declarations") %><br><%= calculator.total_declarations %></span>
+    <span class="govuk-body-s"><%= t("finance.total_not_eligible_for_funding") %><br><%= calculator.not_eligible_declarations %></span>
+  </header>
+
+  <span class="govuk-label--s govuk-visually-hidden">Payment details</span>
+
+  <table class="govuk-table">
+    <thead class="govuk-table__head">
+      <tr class="govuk-table__row">
+        <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.payment_type") %></th>
+        <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.trainees") %></th>
+        <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.payment_per_trainee") %></th>
+        <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.total") %></th>
+      </tr>
+    </thead>
+
+    <tbody class="govuk-table__body">
+      <tr class="gov-table__row">
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= t("finance.output_payment.label") %></td>
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= calculator.total_declarations %></td>
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds calculator.output_payment_per_participant %></td>
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds calculator.output_payment_subtotal %></td>
+      </tr>
+
+      <tr class="gov-table__row">
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= t("finance.service_fee.caption") %></td>
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= calculator.recruitment_target %></td>
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds calculator.service_fees_per_participant %></td>
+        <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds calculator.monthly_service_fees %></td>
+      </tr>
+    </tbody>
+  </table>
+
+  <div class="govuk-table__row govuk-!-padding-bottom-1">
+    <div class="govuk-label--s govuk-table__header--numeric">
+      <%= t("finance.course_total") %>
+      <br>
+      <%= number_to_pounds calculator.course_total %>
+    </div>
+  </div>
+</section>
