--- app/components/finance/npq/payment_overviews/payment_overview_milestone_summary.html.erb
@@ -1,5 +1,0 @@
-<span class="govuk-body-s">
-  <%= declaration_type %>
-  <br>
-  <%= total_participants %>
-</span>
