--- app/components/finance/npq/payment_overviews/payment_overview_milestone_summary.rb
@@ -1,26 +1,0 @@
-# frozen_string_literal: true
-
-module Finance
-  module NPQ
-    module PaymentOverviews
-      class PaymentOverviewMilestoneSummary < BaseComponent
-        include NPQPaymentsHelper
-        attr_reader :total_participants
-
-        def initialize(milestone, total_participants)
-          self.milestone = milestone
-          self.total_participants = total_participants
-        end
-
-        def declaration_type
-          milestone.declaration_type.titleize
-        end
-
-      private
-
-        attr_accessor :milestone
-        attr_writer :total_participants
-      end
-    end
-  end
-end
