--- app/components/finance/npq/payment_overviews/payment_overview_table.html.erb
@@ -1,15 +1,0 @@
-<table class="govuk-table">
-  <thead class="govuk-table__head">
-  <tr class="govuk-table__row">
-    <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.payment_type") %></th>
-    <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.trainees") %></th>
-    <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.payment_per_trainee") %></th>
-    <th scope="col" class="govuk-table__header govuk-body-s"><%= t("finance.total") %></th>
-  </tr>
-  </thead>
-  <tbody class="govuk-table__body">
-  <%= render Finance::NPQ::PaymentOverviews::OutputPaymentRow.new(output_payment, contract, statement, npq_lead_provider) %>
-  <%= render Finance::NPQ::PaymentOverviews::ServiceFeeRow.new(service_fees, contract) %>
-  </tbody>
-</table>
-<%= render Finance::NPQ::PaymentOverviews::PaymentOverviewTotal.new(service_fees, output_payment, npq_lead_provider) %>
