--- app/components/finance/npq/payment_overviews/payment_overview_table_header.html.erb
@@ -1,7 +1,0 @@
-<header class="app-application-card__header">
-  <% milestones.order(name: :asc).each do |milestone| %>
-    <%= render Finance::NPQ::PaymentOverviews::PaymentOverviewMilestoneSummary.new(milestone, total_participants_for(milestone)) %>
-  <% end %>
-  <span class="govuk-body-s"><%= t("finance.total_declarations") %><br><%= total_declarations(contract) %></span>
-  <span class="govuk-body-s"><%= t("finance.total_not_eligible_for_funding") %><br><%= not_eligible_declarations %></span>
-</header>
