--- app/components/finance/npq/payment_overviews/payment_overview_total.html.erb
@@ -1,7 +1,0 @@
-<div class="govuk-table__row govuk-!-padding-bottom-1">
-  <div class="govuk-label--s govuk-table__header--numeric">
-    <%= t("finance.course_total") %>
-    <br>
-    <%= number_to_pounds course_total %>
-  </div>
-</div>
