--- app/components/finance/npq/payment_overviews/service_fee_row.html.erb
@@ -1,6 +1,0 @@
-<tr class="gov-table__row">
-  <td scope="row" class="govuk-table__cell govuk-body-s"><%= t("finance.service_fee.caption") %></td>
-  <td scope="row" class="govuk-table__cell govuk-body-s"><%= recruitment_target %></td>
-  <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds service_fees_per_participant %></td>
-  <td scope="row" class="govuk-table__cell govuk-body-s"><%= number_to_pounds monthly_service_fees %></td>
-</tr>
